+++ src/components/organisms/mobile-menu/index.ts
@@ -1,0 +1,2 @@
+export { MobileMenu as default } from './component';
+export * from './component';
