+++ src/components/organisms/text-with-image/index.ts
@@ -1,0 +1,2 @@
+export { TextWithImage as default } from './component';
+export * from './component';
